+++ lib/ansible/modules/system/setup.py
@@ -118,91 +118,14 @@ EXAMPLES = """
 # Display facts from Windows hosts with custom facts stored in C(C:\\custom_facts).
 # ansible windows -m setup -a "fact_path='c:\\custom_facts'"
 """
-import fnmatch
-import sys
 
 # import module snippets
 from ansible.module_utils.basic import AnsibleModule
 
-from ansible.module_utils.facts import collector
 from ansible.module_utils.facts.namespace import PrefixFactNamespace
+from ansible.module_utils.facts import ansible_collector
 
 from ansible.module_utils.facts import default_collectors
-
-
-# This is the main entry point for setup.py facts.py.
-# FIXME: This is coupled to AnsibleModule (it assumes module.params has keys 'gather_subset',
-#        'gather_timeout', 'filter' instead of passing those are args or oblique ds
-#        module is passed in and self.module.misc_AnsibleModule_methods
-#        are used, so hard to decouple.
-
-class AnsibleFactCollector(collector.BaseFactCollector):
-    '''A FactCollector that returns results under 'ansible_facts' top level key.
-
-       Has a 'from_gather_subset() constructor that populates collectors based on a
-       gather_subset specifier.'''
-
-    def __init__(self, collectors=None, namespace=None, filter_spec=None):
-
-        super(AnsibleFactCollector, self).__init__(collectors=collectors,
-                                                   namespace=namespace)
-
-        self.filter_spec = filter_spec
-
-    def _filter(self, facts_dict, filter_spec):
-        # assume a filter_spec='' is equilv to filter_spec='*'
-        if not filter_spec or filter_spec == '*':
-            return facts_dict
-
-        return [(x, y) for x, y in facts_dict.items() if fnmatch.fnmatch(x, filter_spec)]
-
-    def collect(self, module=None, collected_facts=None):
-        collected_facts = collected_facts or {}
-
-        facts_dict = {}
-        facts_dict['ansible_facts'] = {}
-
-        for collector_obj in self.collectors:
-            info_dict = {}
-
-            # shallow copy of the accumulated collected facts to pass to each collector
-            # for reference.
-            collected_facts.update(facts_dict['ansible_facts'].copy())
-
-            try:
-
-                # Note: this collects with namespaces, so collected_facts also includes namespaces
-                info_dict = collector_obj.collect_with_namespace(module=module,
-                                                                 collected_facts=collected_facts)
-            except Exception as e:
-                sys.stderr.write(repr(e))
-                sys.stderr.write('\n')
-
-            # filtered_info_dict = self._filter(info_dict, self.filter_spec)
-            # NOTE: If we want complicated fact dict merging, this is where it would hook in
-            facts_dict['ansible_facts'].update(self._filter(info_dict, self.filter_spec))
-
-        # TODO: this may be best place to apply fact 'filters' as well. They
-        #       are currently ignored -akl
-        return facts_dict
-
-
-class CollectorMetaDataCollector(collector.BaseFactCollector):
-    '''Collector that provides a facts with the gather_subset metadata.'''
-
-    name = 'gather_subset'
-    _fact_ids = set([])
-
-    def __init__(self, collectors=None, namespace=None, gather_subset=None, module_setup=None):
-        super(CollectorMetaDataCollector, self).__init__(collectors, namespace)
-        self.gather_subset = gather_subset
-        self.module_setup = module_setup
-
-    def collect(self, module=None, collected_facts=None):
-        meta_facts = {'gather_subset': self.gather_subset}
-        if self.module_setup:
-            meta_facts['module_setup'] = self.module_setup
-        return meta_facts
 
 
 def main():
@@ -231,38 +154,21 @@ def main():
 
     all_collector_classes = default_collectors.collectors
 
-    collector_classes = \
-        collector.collector_classes_from_gather_subset(
-            all_collector_classes=all_collector_classes,
-            minimal_gather_subset=minimal_gather_subset,
-            gather_subset=gather_subset,
-            gather_timeout=gather_timeout)
-
-    # print('collector_classes: %s' % pprint.pformat(collector_classes))
-
+    # rename namespace_name to root_key?
     namespace = PrefixFactNamespace(namespace_name='ansible',
                                     prefix='ansible_')
 
-    collectors = []
-    for collector_class in collector_classes:
-        collector_obj = collector_class(namespace=namespace)
-        collectors.append(collector_obj)
-
-    # Add a collector that knows what gather_subset we used so it it can provide a fact
-    collector_meta_data_collector = \
-        CollectorMetaDataCollector(gather_subset=gather_subset,
-                                   module_setup=True)
-    collectors.append(collector_meta_data_collector)
-
-    # print('collectors: %s' % pprint.pformat(collectors))
-
     fact_collector = \
-        AnsibleFactCollector(collectors=collectors,
-                             filter_spec=filter_spec)
+        ansible_collector.get_ansible_collector(all_collector_classes=all_collector_classes,
+                                                namespace=namespace,
+                                                filter_spec=filter_spec,
+                                                gather_subset=gather_subset,
+                                                gather_timeout=gather_timeout,
+                                                minimal_gather_subset=minimal_gather_subset)
 
     facts_dict = fact_collector.collect(module=module)
 
-    module.exit_json(**facts_dict)
+    module.exit_json(ansible_facts=facts_dict)
 
 
 if __name__ == '__main__':
